backend/requirements.txt
flask
flask-cors
mysql-connector-python
python-dotenv
langchain
langchain-community
langchain-chroma
langchain-core
langchain-groq
chromadb